Correcion Pantallas Personal Pasantes
Faltan cambios
